Rename getRequestOptions
